Clang/MSVC warning fixes

+++ core/src/shaders/Shader.cpp
@@ -33,9 +33,9 @@ Shader& Shader::AddStage(vk::ShaderStageFlagBits shader_stage_flags, const std::
     shader_create_info.codeSize = spv.size();
     shader_create_info.pCode = spv.data();
     shader_create_info.pName = "main";
-    shader_create_info.setLayoutCount = m_set_layouts.size();
+    shader_create_info.setLayoutCount = (uint32_t)m_set_layouts.size();
     shader_create_info.pSetLayouts = m_set_layouts.data();
-    shader_create_info.pushConstantRangeCount = m_push_constant_ranges.size();
+    shader_create_info.pushConstantRangeCount = (uint32_t)m_push_constant_ranges.size();
     shader_create_info.pPushConstantRanges = m_push_constant_ranges.data();
 
     m_shader_stages.emplace_back(shader_stage_flags);
@@ -47,13 +47,13 @@ vk::PipelineLayout Shader::GetPipelineLayout() {
 }
 
 void Shader::Bind(vk::CommandBuffer& cmd) {
-    cmd.bindShadersEXT(m_shaders.size(), m_shader_stages.data(), m_shaders.data());
+    cmd.bindShadersEXT((uint32_t)m_shaders.size(), m_shader_stages.data(), m_shaders.data());
 }
 
 void Shader::Build() {
     m_shaders.resize(m_shader_create_infos.size());
     OGFX_VK_CHECK(VkContext::GetLogicalDevice().device.createShadersEXT(
-        m_shader_create_infos.size(), m_shader_create_infos.data(), nullptr, m_shaders.data()));
+        (uint32_t)m_shader_create_infos.size(), m_shader_create_infos.data(), nullptr, m_shaders.data()));
 
     m_spv_storage.clear();
     m_shader_create_infos.clear();
@@ -86,16 +86,16 @@ void Shader::ReflectPipelineLayout(const std::vector<std::byte>& spv) {
     for (const spirv_cross::Resource& resource : resources.push_constant_buffers) {
         vk::PushConstantRange& push_constant_range = m_push_constant_ranges.emplace_back();
         push_constant_range.offset = compiler.get_decoration(resource.id, spv::DecorationOffset);
-        push_constant_range.size = compiler.get_declared_struct_size(compiler.get_type(resource.type_id));
+        push_constant_range.size = (uint32_t)compiler.get_declared_struct_size(compiler.get_type(resource.type_id));
         push_constant_range.stageFlags = vk::ShaderStageFlagBits::eAll; // TODO: Make this more specific
     }
 
     m_set_layouts.resize(sets.size());
 
-    int idx = 0;
+    uint32_t idx = 0;
     for (const auto& set: sets | std::views::values) {
         vk::DescriptorSetLayoutCreateInfo info{};
-        info.bindingCount = set.bindings.size();
+        info.bindingCount = (uint32_t)set.bindings.size();
         info.pBindings = set.bindings.data();
         info.flags = vk::DescriptorSetLayoutCreateFlagBits::ePushDescriptor;
 
@@ -104,9 +104,9 @@ void Shader::ReflectPipelineLayout(const std::vector<std::byte>& spv) {
     }
 
     vk::PipelineLayoutCreateInfo pipeline_layout_create_info{};
-    pipeline_layout_create_info.setLayoutCount = m_set_layouts.size();
+    pipeline_layout_create_info.setLayoutCount = (uint32_t)m_set_layouts.size();
     pipeline_layout_create_info.pSetLayouts = m_set_layouts.data();
-    pipeline_layout_create_info.pushConstantRangeCount = m_push_constant_ranges.size();
+    pipeline_layout_create_info.pushConstantRangeCount = (uint32_t)m_push_constant_ranges.size();
     pipeline_layout_create_info.pPushConstantRanges = m_push_constant_ranges.data();
 
     m_pipeline_layout = ogfx::VkContext::GetLogicalDevice().device.createPipelineLayoutUnique(pipeline_layout_create_info);
